+++ src/dmLib/__init__.py
@@ -8,7 +8,7 @@ uncertainty modelling
 
 """
 
-__version__ = '0.4.4'
+__version__ = '0.4.6'
 __all__=['triangularFunc', 'fuzzySet', 'fuzzyRule','fuzzySystem', 'gaussianFunc', 'uniformFunc', 'Distribution',
     'VisualizeDist', 'compute_cdf', 'Design', 'MarginNode','InputSpec', 'FixedParam', 'DesignParam', 'Behaviour', 
     'Performance', 'ImpactMatrix', 'MarginNetwork']
@@ -16,4 +16,4 @@ __all__=['triangularFunc', 'fuzzySet', 'fuzzyRule','fuzzySystem', 'gaussianFunc'
 from .fuzzyLib import triangularFunc, fuzzySet, fuzzyRule, fuzzySystem
 from .uncertaintyLib import gaussianFunc, uniformFunc, Distribution, VisualizeDist, compute_cdf
 from .DOELib import Design
-from .designMarginsLib import InputSpec, FixedParam, DesignParam, Behaviour, MarginNode, Performance, ImpactMatrix, MarginNetwork
+from .designMarginsLib import InputSpec, FixedParam, DesignParam, Behaviour, MarginNode, Performance, ImpactMatrix, MarginNetwork, nearest